rename files to ensure consistent lower casing
