+++ tslib/src/functions/check-allocator/check-allocator.spec.ts
@@ -2,10 +2,10 @@ import * as uuid from 'uuid'
 import moment from 'moment'
 import config from '../../config'
 import { SchoolCheckAllocationService } from './check-allocator'
-import { IPupil, ISchoolAllocation } from './Models'
-import { ICheckAllocationDataService } from './CheckAllocationDataService'
+import { IPupil, ISchoolAllocation } from './models'
+import { ICheckAllocationDataService } from './check-allocation.data.service'
 import { IDateTimeService } from '../../common/DateTimeService'
-import { IPupilAllocationService } from './PupilAllocationService'
+import { IPupilAllocationService } from './pupil-allocation.service'
 import { IRedisService } from '../../caching/redis-service'
 import { ConsoleLogger } from '../../common/ILogger'
