+++ tslib/src/functions/check-allocator/check-window.data.service.ts
