+++ tslib/src/functions/check-validator/check-validator.v1.spec.ts
@@ -1,6 +1,6 @@
 import * as CheckValidator from './check-validator.v1'
 import { IAsyncTableService } from '../../azure/storage-helper'
-import { ValidateCheckMessageV1, ReceivedCheck, MarkCheckMessageV1 } from '../../schemas/Models'
+import { ValidateCheckMessageV1, ReceivedCheck, MarkCheckMessageV1 } from '../../schemas/models'
 import { ILogger } from '../../common/ILogger'
 import checkSchema from '../../schemas/complete-check.v1.json'
 import { ICompressionService } from '../../common/compression-service'
